requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
idna==3.3
iniconfig==1.1.1
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
pytest-django==4.5.2
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
